+++ crates/nostr/src/event/tag.rs
@@ -1595,21 +1595,31 @@ mod tests {
         .unwrap();
         assert_eq!(
             t.content(),
-            Some(GenericTagValue::Pubkey(
+            Some(GenericTagValue::Fixed32Bytes(
                 PublicKey::from_hex(
                     "f86c44a2de95d9149b51c6a29afeabba264c18e2fa7c49de93424a0c56947785"
                 )
                 .unwrap()
+                .to_bytes()
             ))
         );
 
-        // TODO: fix this
-        // // Test extract event ID
-        // let t: Tag = Tag::parse(&["custom-e", "2be17aa3031bdcb006f0fce80c146dea9c1c0268b0af2398bb673365c6444d45"]).unwrap();
-        // assert_eq!(
-        //     t.content(),
-        //     Some(GenericTagValue::EventId(EventId::from_hex("2be17aa3031bdcb006f0fce80c146dea9c1c0268b0af2398bb673365c6444d45").unwrap()))
-        // );
+        // Test extract event ID
+        let t: Tag = Tag::parse(&[
+            "custom-e",
+            "2be17aa3031bdcb006f0fce80c146dea9c1c0268b0af2398bb673365c6444d45",
+        ])
+        .unwrap();
+        assert_eq!(
+            t.content(),
+            Some(GenericTagValue::Fixed32Bytes(
+                EventId::from_hex(
+                    "2be17aa3031bdcb006f0fce80c146dea9c1c0268b0af2398bb673365c6444d45"
+                )
+                .unwrap()
+                .to_bytes()
+            ))
+        );
     }
 
     #[test]