+++ CHANGELOG.md
@@ -23,7 +23,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-* `Tag::content` return always `None` when `Tag::Generic` ([Yuki Kishimoto])
+* nostr: `Tag::content` return always `None` when `Tag::Generic` ([Yuki Kishimoto])
+* nostr: fix `GenericTagValue` deserialization ([Yuki Kishimoto])
 
 ### Removed